Move function:NormalizeDomain to netutils module
